requirements.txt
# Volatility Path States - Requirements
# =====================================
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Statistical analysis
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Data access
yfinance>=0.2.28
pandas-datareader>=0.10.0
wrds>=3.1.6  # Optional: for WRDS/CRSP data access

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Progress bars and utilities
tqdm>=4.65.0
joblib>=1.3.0

# Date/time handling
python-dateutil>=2.8.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.25.0

# Documentation (optional)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Type checking (optional)
mypy>=1.5.0
pandas-stubs>=2.0.0
